Main 1 terminado

+++ KPIs/Vacuna.java
@@ -1,5 +1,5 @@
 package KPIs;
 
 public interface Vacuna {
-    float calcularPorcentajeVacunacion();
+    double calcularPorcentajeVacunacion();
 }
